requirements.txt
fastapi
uvicorn
python-multipart
streamlit
PyMuPDF
python-docx
faiss-cpu
sentence-transformers
transformers
llama-cpp-python  # For running local LLM like Mistral or TinyLLaMA
chromadb
sentence-transformers
python-docx
pdfplumber
transformers
torch
